// src/app/models/api.model.ts
export interface Pokemons {
  count: number;
  next: string;
  previous: any;
  results: Result[];
}

export interface Result {
  name: string;
  url: string;
}